java-github-actions-playground-8 project 連携 - gradle タスクを起動して、見積もりをどこかに保存する refactor
