+++ buildSrc/src/main/java/com/example/gradle/aws/AwsDynamoDbClientConfig.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.gradle
+package com.example.gradle.aws
 
 import com.amazonaws.auth.AWSCredentialsProvider
 import com.amazonaws.auth.AWSStaticCredentialsProvider
