+++ buildSrc/src/main/java/com/example/gradle/db/create-tables.kt
@@ -19,6 +19,8 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder
 import com.amazonaws.services.dynamodbv2.model.*
 import com.example.gradle.*
+import com.example.gradle.aws.CredentialConfig
+import com.example.gradle.aws.EndpointConfig
 
 class DynamoDbCreateTable(
     private val client: AmazonDynamoDB,
